update : made some file restructuring and naming
