+++ src/interfaces/cachie.interface.ts
