+++ src/middlewares/rate-limiter.middleware.ts
@@ -1,5 +1,5 @@
 import { Request, Response, NextFunction } from "express";
-import logger from "../logger";
+import logger from "../utils/logger.util";
 
 interface RateLimitData {
   count: number;
