+++ src/server.ts
@@ -3,7 +3,7 @@ import { initialize } from "express-openapi";
 import path from "path";
 import router from "./routes/cachie.route";
 import "dotenv/config";
-import logger from "./logger";
+import logger from "./utils/logger.util";
 import swaggerUi from "swagger-ui-express";
 import fs from "fs";
 import yaml from "js-yaml";
